Modified reads to accommodate the new directly evaluated vzsum. piccompost modified accordingly.
Added some fgencol to calculate collisional contributions to the momentum
based mostly on Leonardo's fgen but with some additional tweaks. Notably
to print out a comparison between the vzsum and the equivalent vr vt calc.

forcehist just plots the force trends for a visual inspection of whether
we seem to be converged.