This commit is mostly for the copyright/versioning tools. However, all .f and .F files are touched by the versioning process. That is why there's a new version of them all. The versions ought not to be functionally different from prior versions, but if they are, that implies I did something wrong that should be back-tracked.
At any rate the changes _ought_ to be just in the headers.
We now store in the CVS repository files without the copyright/version
information at the top.

The script copyattach.sh attaches the information and versions the files.
By default it uses the cvs current version of sceptic.F as the version.
That can be changed by arguments to the script.
If the information is attached, the files ought not to be committed back
to the repository in that state. Instead, the script copyremove.sh ought
to be run to remove the information before anything is copied back.
If you forget, it should not be a big problem, someone will have to
get the files out of the repository, copyremove them and commit them back.

make sceptic.tar.gz causes the files to be copyattached, tarred, then
copyremoved. The result ought to be that the files are the same as they
were before, but we have a tar file of the whole system with copyright
files.